restructured files. routes added to protected or public in middleware.
