+++ src/ middleware.ts
@@ -2,8 +2,8 @@ import { NextRequest, NextResponse } from "next/server";
 import { cookies } from "next/headers";
 import { decrypt } from "./app/lib/sessions";
 
-const protectedRoutes = ["/dashboard"];
-const publicRoutes = ["/login"];
+const protectedRoutes = ["/dashboard", "day", "mistakes", "setup", "trade", "triggers"];
+const publicRoutes = ["/login", "register"];
 
 export default async function middleware(req: NextRequest) {
   const path = req.nextUrl.pathname;
