+++ src/app/dashboard/page.tsx
@@ -1,0 +1,6 @@
+
+export default function DashboardPage() {
+  return (
+    <h1>Dashboard Page</h1>
+  );
+}
